Move Kt + Wt sum to message scheduler.
Summary:

This commit moves Kt + Wt operation from compression function
to message scheduler. Both Wt and Kt are available and
are independent of the other operands we can do the sum on
message scheduler. This may improve performance since we
already have the operands avaliable so we don't need to load
those operands from memory on compression phase and we can
do message scheduler in parallel (using vector instructions)
we can sum Kt in parallel too.

This commit only do this to the version who uses gcc builtin
for now. Still need to
- do that on asm inline version;
- check if there's a performance improvement or issue.