Fix undefined behaviour in tests.
Summary:

This commit fixes the undefined behaviour that occours in
test when compile with -fsanitize=address,undefined. When U
or UL suffix are not used, the compiler uses the first type
that can contains the constant. Since we pass constants to
macros that 'fits' on int and long int hence, we got 'cannot
be represented' errors because some shifts overflow those
(signed) types. Use unsigned types instead.

Fixes issue #5